error model

+++ util.py
@@ -42,7 +42,7 @@ def read_color_data_set(datadir):
 
     for i in range(cat_len):
         img_g = cv2.imread(os.path.join(datadir, x_list[i]), cv2.IMREAD_GRAYSCALE)
-        img_g = cv2.resize(img_g, dsize=(256, 256), interpolation=cv2.INTER_CUBIC).astype(np.float32)
+        img_g = cv2.resize(img_g, dsize=(256, 256), interpolation=cv2.INTER_CUBIC)
         img_g = img_g[:, :, np.newaxis]
         x_features[i] = img_g
         img = cv2.imread(os.path.join(datadir, y_list[i]))
